Lookup conditions using Yahoo Weather API
If new -location flag is provided, the current conditions as returned
by Yahoo's Weather API is used to determine -temp, -conditions and
-wind.

If -time is not specified or has the value 'current', determine an
appropriate value based on the current time.

With these changes, only -location, -gender, -intensity and -feel are
needed to get an answer.